// mobile_app/components/followup/interface.ts
export interface Symptome {
  title: string;
  month: Month;
  slug: string;
  status: string;
  code: string;
  currentMonth?: number;
  phoneNumber?: string;
  secondPhoneNumber?: string;
  urgencyModuleText?: string;
  urgencyPageText?: string;
  urgencyPageMainTitle?: string;
  urgencyPageSubTitle?: string;
}

export interface Meetings {
  title: string;
  code: string;
  isMandatory: boolean;
  months?: Month[];
  monthNumber?: number;
  maxMonth?: number;
  hasMoreInfo?: boolean;
  meeting_info?: MeetingInfo;
}

export interface MeetingInfo {
  title: string;
  description: string;
  img_slug: string;
}

export interface Month {
  title: string;
  description: string;
  monthNumber: number;
  meetings: Meetings[];
  symptoms: Symptome[];
}